Moving `XFCCError` into `xfcc_errors.go`.
